+++ src/lib/components/Link.svelte
@@ -37,9 +37,8 @@
 	}
 
 	a:focus {
-		outline: var(--focus-color) solid var(--thick-border-width);
-		border-radius: var(--roundedness);
-		outline-offset: var(--border-width);
+		outline: none;
+		color: var(--focus-color);
 	}
 
 	a[aria-current] {